Android: Now using an activity for Preferences instead of a fragment, since Android doesn't support fragments in fragments (i.e. Fragment with a ViewPager). Changed the way views are loaded with MobileNavigationManager to support Android intents (i.e. not having a reference to a view you create). Extremely infuriating how Android manages view creation compared to iOS.
Related to issue #406.

+++ MPfm/MPfm.Android/Classes/Fragments/UpdateLibraryFragment.cs
@@ -66,7 +66,7 @@ namespace MPfm.Android.Classes.Fragments
         {
             base.OnCreate(savedInstanceState);
             Cancelable = false;
-            SetStyle(DialogFragmentStyle.Normal, Resource.Style.UpdateLibraryTheme);
+            SetStyle((int)DialogFragmentStyle.Normal, (int)Resource.Style.UpdateLibraryTheme);
         }
 
         public override void OnStart()
